requirements.txt
python-binance==1.0.19
pandas==2.2.2
python-dotenv==1.0.1
